requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
Django==1.9.1
django-recommends==0.3.1
djangorestframework==3.3.3
kombu==3.0.33
python-dateutil==2.4.2
python-recsys==0.2
pytz==2015.7
selenium==2.49.2
six==1.10.0
wheel==0.26.0
